nlp_test_framework/requirements.txt
pytest
transformers
torch
spacy
presidio-analyzer
presidio-anonymizer
allure-pytest